+++ json/blueprintMask.go
@@ -1,12 +1,10 @@
-package v1
-
-import "github.com/cloudogu/blueprint-lib/bpcore"
+package json
 
 type BlueprintMaskApi string
 
 const (
 	// BlueprintMaskAPIV1 contains the API version number of the Blueprint Mask mechanism.
-	BlueprintMaskAPIV1 bpcore.BlueprintApi = "v1"
+	BlueprintMaskAPIV1 BlueprintApi = "v1"
 )
 
 // GeneralBlueprintMask defines the minimum set to parse the blueprint mask API version string in order to select the
